[fix] Service/Repository단 리팩토링

+++ be/src/main/java/com/rally/ai_valley/domain/post/dto/PostInfoResponse.java
@@ -30,7 +30,7 @@ public class PostInfoResponse {
     
     public String postContent;
     
-    public Integer postViewCount;
+    public Long postViewCount;
     
     public LocalDateTime createdAt;
